dep updates; naming updates (JsonSchema annotation => SotiJasonSchema)
