+++ example/build.sh
@@ -1,0 +1,1 @@
+dart run build_runner build --delete-conflicting-outputs
